requirements.txt
pandas 
numpy 
ta 
lightgbm 
streamlit 
plotly 
ccxt 
pyarrow 
transformers 
accelerate 
einops 
chronos-forecasting
fastapi
uvicorn